chore: unwrap lib folder in app
