+++ apps/web/src/app/[locale]/search/search-page.tsx
@@ -11,7 +11,7 @@ import SearchForm from '@/components/SearchForm';
 import { ArticleSearchResult } from '@/features/content';
 import { AppArticleService } from '@/features/content/services/AppArticleService';
 import { Link } from '@/i18n/navigation';
-import { useExperienceFlagsStore } from '@/lib/state/experience_flags';
+import { useExperienceFlagsStore } from '@/state/experience_flags';
 
 type Result = {
   query: string;
