+++ apps/web/src/app/[locale]/user/login/login-page.tsx
@@ -13,7 +13,7 @@ import { SubmitHandler, useForm } from 'react-hook-form';
 
 import { LoginFormInputs } from '@/features/auth';
 import { Link } from '@/i18n/navigation';
-import { EMAIL_PATTERN } from '@/lib/utils/validator';
+import { EMAIL_PATTERN } from '@/utils/validator';
 
 export function LoginPage() {
   const t = useTranslations();
