+++ apps/web/src/app/[locale]/user/password-reminder/password-reminder-page.tsx
@@ -17,7 +17,7 @@ import {
 
 import { PasswordReminderFormInputs } from '@/features/auth';
 import { Link } from '@/i18n/navigation';
-import { EMAIL_PATTERN } from '@/lib/utils/validator';
+import { EMAIL_PATTERN } from '@/utils/validator';
 
 export function PasswordReminderPage() {
   const t = useTranslations();
