+++ apps/web/src/features/browser_core/hooks/useDisableContextMenu.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from 'next-intl';
 import { useCallback, useEffect } from 'react';
 
-import { useExperienceFlagsStore } from '@/lib/state/experience_flags';
+import { useExperienceFlagsStore } from '@/state/experience_flags';
 
 const useDisableContextMenu = () => {
   const t = useTranslations();
