+++ apps/web/src/features/chat_bubble/hooks/useChatBubbleHistory.ts
@@ -2,10 +2,10 @@ import { useMessages, useTranslations } from 'next-intl';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { HistoryItem, HistoryItemOwner } from '@/features/chat_bubble';
-import useAudio from '@/lib/hooks/useAudio';
-import useSendNotification from '@/lib/hooks/useSendNotification';
-import { useRuntimeStore } from '@/lib/state/runtime';
-import { useUserPreferencesStore } from '@/lib/state/user_preferences';
+import useAudio from '@/hooks/useAudio';
+import useSendNotification from '@/hooks/useSendNotification';
+import { useRuntimeStore } from '@/state/runtime';
+import { useUserPreferencesStore } from '@/state/user_preferences';
 
 const useChatBubbleHistory = () => {
   const t = useTranslations();
