+++ apps/web/src/features/newsletter/NewsletterModalExperienceHost.tsx
@@ -4,9 +4,9 @@ import { FunctionComponent, useEffect, useState } from 'react';
 
 import NewsletterModal from './components/NewsletterModal';
 
-import useScrollDistanceTrigger from '@/lib/hooks/useScrollDistanceTrigger';
-import { useExperienceFlagsStore } from '@/lib/state/experience_flags';
-import { useRuntimeStore } from '@/lib/state/runtime';
+import useScrollDistanceTrigger from '@/hooks/useScrollDistanceTrigger';
+import { useExperienceFlagsStore } from '@/state/experience_flags';
+import { useRuntimeStore } from '@/state/runtime';
 
 export type NewsletterModalExperienceHostProps = {
   scrollDistanceTrigger?: number;
