+++ apps/web/src/features/settings/components/PreferencesSettings.tsx
@@ -5,8 +5,8 @@ import { useTranslations } from 'next-intl';
 import { useTheme } from 'next-themes';
 import { FunctionComponent } from 'react';
 
-import { useRuntimeStore } from '@/lib/state/runtime';
-import { useUserPreferencesStore } from '@/lib/state/user_preferences';
+import { useRuntimeStore } from '@/state/runtime';
+import { useUserPreferencesStore } from '@/state/user_preferences';
 
 const PreferencesSettings: FunctionComponent = () => {
   const preference = useUserPreferencesStore();
