+++ apps/web/src/hooks/useDocumentVisibilityListener.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect } from 'react';
 
-import { useRuntimeStore } from '@/lib/state/runtime';
+import { useRuntimeStore } from '@/state/runtime';
 
 /**
  * This will mesaure how long the webpage has been in focus and report it to
