+++ apps/web/src/hooks/useNavigationStats.ts
@@ -1,7 +1,7 @@
 import { usePathname } from 'next/navigation';
 import { useEffect, useState } from 'react';
 
-import { useRuntimeStore } from '@/lib/state/runtime';
+import { useRuntimeStore } from '@/state/runtime';
 
 const useNavigationStats = () => {
   const incrementNavigationCount = useRuntimeStore(
