+++ apps/web/src/providers/RootProviderContainer.tsx
@@ -3,8 +3,8 @@
 import { ThemeProvider } from 'next-themes';
 import { FunctionComponent, PropsWithChildren } from 'react';
 
-import ExperienceProvider from '@/lib/providers/ExperienceProvider';
-import { RootPortalProvider } from '@/lib/providers/RootPortalProvider';
+import ExperienceProvider from '@/providers/ExperienceProvider';
+import { RootPortalProvider } from '@/providers/RootPortalProvider';
 
 const RootProviderContainer: FunctionComponent<PropsWithChildren> = ({
   children,
